Update form for caloric calculator, add UI mod script for dynamic select options depending on chosen unit of measurement for height

+++ formulas.py
@@ -7,13 +7,14 @@ def basal_metabolic_rate(height, weight, age, sex):
         return 655 + (4.35 * weight) + (4.7 * weight) - (4.7 * age)
 
 def harris_benedict_calculation(activity_level, bmr):
+    # calculates recommended caloric intake using the harris benedict equation
     if activity_level == 'inactive':
-        return bmr * 1.2
+        return int(bmr * 1.2)
     elif activity_level == 'somewhat_active':
-        return bmr * 1.375
+        return int(bmr * 1.375)
     elif activity_level == 'active':
-        return bmr * 1.55
+        return int(bmr * 1.55)
     elif activity_level == 'athlete':
-        return bmr * 1.725
+        return int(bmr * 1.725)
     else:
         return print("error: enter a valid return type")
